Rename manual feature package to debug
