+++ WordPress/src/main/res/layout/manual_feature_config_activity.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/fragment_container"
-        android:name="org.wordpress.android.ui.manual.ManualFeatureConfigFragment"
+        android:name="org.wordpress.android.ui.debug.ManualFeatureConfigFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
